Extract stack rules
