+++ engine/src/main/kotlin/world/gregs/voidps/engine/inv/stack/ItemDependentStack.kt
@@ -1,0 +1,12 @@
+package world.gregs.voidps.engine.inv.stack
+
+import world.gregs.voidps.engine.data.definition.ItemDefinitions
+
+/**
+ * Checks individual items are stackable
+ */
+class ItemDependentStack(
+    private val definitions: ItemDefinitions
+) : ItemStackingRule {
+    override fun stackable(id: String) = definitions.get(id).stackable == 1
+}
