+++ engine/src/main/kotlin/world/gregs/voidps/engine/inv/stack/NeverStack.kt
@@ -1,0 +1,8 @@
+package world.gregs.voidps.engine.inv.stack
+
+/**
+ * Forces all items, even stackable items, to be unstacked
+ */
+object NeverStack : ItemStackingRule {
+    override fun stackable(id: String) = false
+}
